+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to the esi-service project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!--## [Unreleased]-->
+## [Unreleased]
+### Added
+- Ability to do a raw request to the ESI with custom config options, this request will not use caching and return the AxiosResponse object.
+- Full headers to cached data, this will increase the cache size, but allows for custom interaction with headers.
+
 ## [3.2.2] - 2019-08-26
 ### Changed
 - `expiry` and `etag` values will be removed from cached requests when they are not in the response when the cache updates.